chore: remove chaindata requirement from getFrontendTypeRegistry function
