+++ apps/extension/src/ui/util/getFrontendTypeRegistry.ts
@@ -1,31 +1,25 @@
 import { typesBundle } from "@polkadot/apps-config/api"
 import { Metadata, TypeRegistry } from "@polkadot/types"
 import { getSpecAlias, getSpecTypes } from "@polkadot/types-known/util"
-import { hexToNumber, isHex } from "@polkadot/util"
+import { hexToNumber } from "@polkadot/util"
 import { HexString } from "@polkadot/util/types"
 
 import { Chain } from "@extension/core"
 import { getMetadataFromDef, getMetadataRpcFromDef, log } from "@extension/shared"
 import { api } from "@ui/api"
-import { chaindataProvider } from "@ui/domains/Chains/chaindataProvider"
 
 /**
  * do not reuse getTypeRegistry because we're on frontend, we need to leverage backend's metadata cache
  */
 export const getFrontendTypeRegistry = async (
-  chainIdOrHash: string,
+  chain?: Chain,
   specVersion?: number | string,
   blockHash?: string,
   signedExtensions?: string[],
 ) => {
   const registry = new TypeRegistry()
 
-  // TODO remove type override once chaindata-provider is fixed
-  const chain = (await (isHex(chainIdOrHash)
-    ? chaindataProvider.chainByGenesisHash(chainIdOrHash)
-    : chaindataProvider.chainById(chainIdOrHash))) as Chain | null
-
-  const genesisHash = (isHex(chainIdOrHash) ? chainIdOrHash : chain?.genesisHash) as HexString
+  const genesisHash = chain?.genesisHash as HexString
 
   // register typesBundle in registry for legacy (pre metadata v14) chains
   if (typesBundle.spec && chain?.specName && typesBundle.spec[chain.specName]) {
